fix e2e tests to use sqlite

+++ test/setup-e2e.ts
@@ -11,6 +11,11 @@ function generateUniqueDatabaseURL(schemaId: string) {
     throw new Error('Please provider a DATABASE_URL environment variable')
   }
 
+  if (process.env.DATABASE_URL.startsWith('file:')) {
+    databases.push(schemaId)
+    return `file:./test/${schemaId}.db`
+  }
+
   const url = new URL(process.env.DATABASE_URL)
 
   url.searchParams.set('schema', schemaId)
@@ -19,6 +24,7 @@ function generateUniqueDatabaseURL(schemaId: string) {
 }
 
 const schemaId = randomUUID()
+const databases: string[] = []
 
 beforeAll(async () => {
   const databaseURL = generateUniqueDatabaseURL(schemaId)
@@ -29,6 +35,13 @@ beforeAll(async () => {
 })
 
 afterAll(async () => {
-  await prisma.$executeRawUnsafe(`DROP SCHEMA IF EXISTS "${schemaId}" CASCADE`)
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  if (!process.env.DATABASE_URL!.startsWith('file:')) {
+    await prisma.$executeRawUnsafe(`DROP SCHEMA IF EXISTS "${schemaId}" CASCADE`)
+  }
   await prisma.$disconnect()
-})
+  databases.forEach((database) => {
+    execSync(`del /s /q prisma\\test\\${database}.db`)
+    console.log(`Clearing database ${database}`)
+  })
+}, 10000)
